Add user authentication

+++ backend/server.js
@@ -1,6 +1,39 @@
 const express = require("express");
 const cors = require("cors");
 const db = require("./db");
+const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
+function verifyToken(req, res, next) {
+
+    const authHeader = req.headers.authorization;
+
+    if (!authHeader) {
+        return res.status(401).json({
+            message: "Access denied. No token provided."
+        });
+    }
+
+    const token = authHeader.split(" ")[1];
+
+    try {
+
+        const decoded = jwt.verify(
+            token,
+            process.env.JWT_SECRET
+        );
+
+        req.user = decoded;
+
+        next();
+
+    } catch (error) {
+
+        return res.status(401).json({
+            message: "Invalid or expired token"
+        });
+
+    }
+}
 const app = express();
 app.use(cors());
 app.use(express.json());
@@ -23,9 +56,12 @@ app.get("/", function (req, res) {
     res.send("TaskFlow Backend is running!");
 });
 
-app.get("/api/tasks", function (req, res) {
-
-    db.query("SELECT * FROM tasks", function(error, results) {
+app.get("/api/tasks", verifyToken, function (req, res) {
+const userId = req.user.userId;
+    db.query(
+    "SELECT * FROM tasks WHERE user_id = ?",
+    [userId],
+    function(error, results)  {
 
         if (error) {
             return res.status(500).json({
@@ -46,14 +82,14 @@ app.get("/api/tasks", function (req, res) {
     });
 
 });
-app.put("/api/tasks/:id", function (req, res) {
-
+ app.put("/api/tasks/:id", verifyToken, function (req, res) {
+const userId = req.user.userId;
     const taskId = Number(req.params.id);
 
     db.query(
-        "SELECT * FROM tasks WHERE id = ?",
-        [taskId],
-        function(error, results) {
+    "SELECT * FROM tasks WHERE id = ? AND user_id = ?",
+    [taskId, userId],
+    function(error, results) {
 
             if (error) {
                 return res.status(500).json({
@@ -82,14 +118,14 @@ app.put("/api/tasks/:id", function (req, res) {
                 : Boolean(task.completed);
 
             const sql = `
-                UPDATE tasks
-                SET name = ?, deadline = ?, completed = ?
-                WHERE id = ?
-            `;
+    UPDATE tasks
+    SET name = ?, deadline = ?, completed = ?
+    WHERE id = ? AND user_id = ?
+`;
 
             db.query(
                 sql,
-                [name, deadline, completed, taskId],
+                [name, deadline, completed, taskId,userId],
                 function(error) {
 
                     if (error) {
@@ -112,13 +148,13 @@ app.put("/api/tasks/:id", function (req, res) {
     );
 
 });
-app.delete("/api/tasks/:id", function (req, res) {
+app.delete("/api/tasks/:id",verifyToken, function (req, res) {
 
     const taskId = Number(req.params.id);
-
-    db.query(
-        "DELETE FROM tasks WHERE id = ?",
-        [taskId],
+const userId = req.user.userId;
+   db.query(
+    "DELETE FROM tasks WHERE id = ? AND user_id = ?",
+    [taskId, userId],
         function(error, result) {
 
             if (error) {
@@ -141,17 +177,20 @@ app.delete("/api/tasks/:id", function (req, res) {
     );
 
 });
-app.post("/api/tasks", function (req, res) {
+ app.post("/api/tasks", verifyToken, function (req, res) {
 
     const name = req.body.name;
     const deadline = req.body.deadline;
+    const userId = req.user.userId;
 
-    const sql = `
-        INSERT INTO tasks (name, deadline, completed)
-        VALUES (?, ?, ?)
-    `;
+    
+       const sql = `
+    INSERT INTO tasks (name, deadline, completed, user_id)
+    VALUES (?, ?, ?, ?)
+`;
+    
 
-    db.query(sql, [name, deadline, false], function(error, result) {
+    db.query(sql, [name, deadline, false,userId], function(error, result) {
 
         if (error) {
             return res.status(500).json({
@@ -163,11 +202,132 @@ app.post("/api/tasks", function (req, res) {
             id: result.insertId,
             name: name,
             deadline: deadline,
-            completed: false
+            completed: false,
+               userId: userId
         };
 
         res.json(newTask);
     });
+
+});
+app.post("/api/signup", async function(req, res) {
+
+    const name = req.body.name;
+    const email = req.body.email;
+    const password = req.body.password;
+
+    if (!name || !email || !password) {
+        return res.status(400).json({
+            message: "All fields are required"
+        });
+    }
+
+    try {
+
+        const hashedPassword = await bcrypt.hash(password, 10);
+
+        const sql = `
+            INSERT INTO users (name, email, password)
+            VALUES (?, ?, ?)
+        `;
+
+        db.query(
+            sql,
+            [name, email, hashedPassword],
+            function(error, result) {
+
+                if (error) {
+
+                    if (error.code === "ER_DUP_ENTRY") {
+                        return res.status(400).json({
+                            message: "Email already exists"
+                        });
+                    }
+
+                    return res.status(500).json({
+                        message: "Database error"
+                    });
+                }
+
+                res.status(201).json({
+                    message: "User registered successfully",
+                    userId: result.insertId
+                });
+
+            }
+        );
+
+    } catch (error) {
+
+        res.status(500).json({
+            message: "Server error"
+        });
+
+    }
+
+});
+app.post("/api/login", async function(req, res) {
+
+    const email = req.body.email;
+    const password = req.body.password;
+
+    if (!email || !password) {
+        return res.status(400).json({
+            message: "Email and password are required"
+        });
+    }
+
+    db.query(
+        "SELECT * FROM users WHERE email = ?",
+        [email],
+        async function(error, results) {
+
+            if (error) {
+                return res.status(500).json({
+                    message: "Database error"
+                });
+            }
+
+            if (results.length === 0) {
+                return res.status(401).json({
+                    message: "Invalid email or password"
+                });
+            }
+
+            const user = results[0];
+
+            const passwordMatch = await bcrypt.compare(
+                password,
+                user.password
+            );
+
+            if (!passwordMatch) {
+                return res.status(401).json({
+                    message: "Invalid email or password"
+                });
+            }
+
+            const token = jwt.sign(
+    {
+        userId: user.id,
+        email: user.email
+    },
+    process.env.JWT_SECRET,
+    {
+        expiresIn: "1h"
+    }
+);
+
+res.json({
+    message: "Login successful",
+    token: token,
+    userId: user.id,
+    name: user.name,
+    email: user.email
+});
+
+        }
+    );
 
 });
